==> Code/02_Preprocessing.ipynb <==
import nltk
import ssl

try:
    _create_unverified_https_context = ssl._create_unverified_context
except AttributeError:
    pass
else:
    ssl._create_default_https_context = _create_unverified_https_context

nltk.download()

import pandas as pd
import re
from nltk.tokenize import sent_tokenize

### Read in News Outlets

#Mainstream national
Bild = pd.read_csv('../Outlets/Bild.csv')
FAZ = pd.read_csv('../Outlets/FAZ.csv')
Focus = pd.read_csv('../Outlets/Focus.csv')
HB = pd.read_csv('../Outlets/HB.csv')
Heute = pd.read_csv('../Outlets/Heute.csv')
NTV = pd.read_csv('../Outlets/NTV.csv')
RND = pd.read_csv('../Outlets/RND.csv')
Spiegel = pd.read_csv('../Outlets/Spiegel.csv')
Tagesschau = pd.read_csv('../Outlets/Tagesschau.csv')
TAZ = pd.read_csv('../Outlets/TAZ.csv')
Vice = pd.read_csv('../Outlets/Vice.csv')
Welt = pd.read_csv('../Outlets/Welt.csv')
SZ = pd.read_csv('../Outlets/SZ.csv')

#Regional media
AZ = pd.read_csv('../Outlets/AZ.csv')
BaZ = pd.read_csv('../Outlets/BaZ.csv')
BM = pd.read_csv('../Outlets/BM.csv')
BeZ = pd.read_csv('../Outlets/BeZ.csv')
HM = pd.read_csv('../Outlets/HM.csv')
HAZ = pd.read_csv('../Outlets/HAZ.csv')
HNA = pd.read_csv('../Outlets/HNA.csv')
LVZ = pd.read_csv('../Outlets/LVZ.csv')
Merkur = pd.read_csv('../Outlets/Merkur.csv')
NB = pd.read_csv('../Outlets/NB.csv')
NK = pd.read_csv('../Outlets/NK.csv')
OZ = pd.read_csv('../Outlets/OZ.csv')
RP = pd.read_csv('../Outlets/RP.csv')
StZ = pd.read_csv('../Outlets/SZ.csv')
TA = pd.read_csv('../Outlets/TA.csv')
Tag24 = pd.read_csv('../Outlets/Tag24.csv')
WAZ = pd.read_csv('../Outlets/WAZ.csv')
Weser = pd.read_csv('../Outlets/Weser.csv')

#Alternative right
AdG = pd.read_csv('../Outlets/AdG.csv')
Compact = pd.read_csv('../Outlets/Compact.csv')
DU = pd.read_csv('../Outlets/DU.csv')
ET = pd.read_csv('../Outlets/ET.csv')
FW = pd.read_csv('../Outlets/FW.csv')
Jouwatch = pd.read_csv('../Outlets/Jouwatch.csv')
JF = pd.read_csv('../Outlets/JF.csv')
MMnews = pd.read_csv('../Outlets/Mmnews.csv')
NP = pd.read_csv('../Outlets/NP.csv')
O24 = pd.read_csv('../Outlets/O24.csv')
PI = pd.read_csv('../Outlets/PI.csv')
PTV = pd.read_csv('../Outlets/PTV.csv')
TE = pd.read_csv('../Outlets/TE.csv')
Zuerst = pd.read_csv('../Outlets/Zuerst.csv')

# #Alternative left
Jacobin = pd.read_csv('../Outlets/Jacobin.csv')
JWe = pd.read_csv('../Outlets/JWe.csv')
JWo = pd.read_csv('../Outlets/JWo.csv')
KgK = pd.read_csv('../Outlets/KgK.csv')
KN = pd.read_csv('../Outlets/KN.csv')
ND = pd.read_csv('../Outlets/ND.csv')
RG = pd.read_csv('../Outlets/RG.csv')
TP = pd.read_csv('../Outlets/TP.csv')



### Merge outlets

national = pd.concat([Bild, FAZ, Focus, HB, Heute, NTV, RND, Spiegel, Tagesschau, TAZ, Vice, Welt, SZ])
regional = pd.concat([AZ, BaZ, BM, BeZ, HM, HAZ, HNA, LVZ, Merkur, NB, NK, OZ, RP, StZ, TA, Tag24, WAZ, Weser])
right = pd.concat([AdG, Compact, DU, ET, FW, Jouwatch, JF, MMnews, NP, O24, PI, PTV, TE, Zuerst])
left = pd.concat([Jacobin, JWe, JWo, KgK, KN, ND, RG, TP])


national['outlet']='national'
regional['outlet']='regional'
right['outlet']='right'
left['outlet']='left'

df = pd.concat([national, regional, right, left])

### Copy DF

This step was completed iteratively in order to remove all articles that had already been processed.

somalia = df.copy()

ids_to_remove1 = somalia_masked['_id'].unique().tolist()
ids_to_remove2 = afghanistan_masked['_id'].unique().tolist()
ids_to_remove3 = syrien_masked['_id'].unique().tolist()
ids_to_remove4 = türkei_masked['_id'].unique().tolist()
ids_to_remove5 = ukraine_masked['_id'].unique().tolist()
ids_to_remove6 = flüchtling_masked['_id'].unique().tolist()
ids_to_remove= ids_to_remove1 + ids_to_remove2 + ids_to_remove3 + ids_to_remove4 +ids_to_remove5 + ids_to_remove6 


migrant = df.copy()
migrant = migrant[~migrant['_id'].isin(ids_to_remove)]

migrant.shape

### Filtering Countries

#regular expression pattern
ukr = r'\b(ukraine(?:r|rs|rn|rin|rinnen))\b'
tur = r'\b(türk(?:e|en|in|innen))\b'
syr = r'\b(syre(?:r|rs|rn|rin|rinnen))\b'
som = r'\b(somalie(?:r|rs|rn|rin|rinnen))\b'
afg = r'\b(afghan(?:e|en|in|innen))\b'
flu = r'(\b(?:flüchtling(?:s|e|en|en)?|geflüchtete(?:r|n)?)\b)'
mig = r'\b(migrant(?:en|in|innen)?|immigrant(?:en|in|innen)?|einwander(?:er|ers|ern|in|innen)?|zuwander(?:er|ers|ern|in|innen)?)\b'

pattern = ukr
matches = df['text'].str.extractall(pattern, flags=re.IGNORECASE)

#get unique matches
for col in matches:
    uniquematch = matches[0].unique()


print(uniquematch)

### Extract sections with matches

def extract_sections(article):
    """ 
    1. splits the articles into sections
    2. extracts sections based on the pattern defined in the previous cell
    3. when a section is extracted and it is shorter or equals 3 sentences then the previous section if there is one and it hasn't been 
    extracted already otherwise it'll extract the next section 
    """
    # Compile the regular expression pattern
    compiled_pattern = re.compile(pattern, re.IGNORECASE)
    
    # Split the article into sections based on two newline characters
    sections = article.split('\n\n')
    
    # Keep track of which sections have been extracted
    extracted_flags = [False] * len(sections)
    
    extracted_groups = []
    for i, section in enumerate(sections):
        if not extracted_flags[i] and compiled_pattern.search(section):
            current_group = []
            sentences = sent_tokenize(section, language='german')
            num_sentences = len(sentences)
            if num_sentences >= 3:
                current_group.append(section)
                extracted_flags[i] = True
            else:
                # Extract the previous section if it exists and hasn't been extracted
                if i > 0 and not extracted_flags[i-1]:
                    current_group.append(sections[i-1])
                    extracted_flags[i-1] = True
                current_group.append(section)
                extracted_flags[i] = True
                # Extract the next section if it exists and hasn't been extracted
                if i < len(sections)-1 and not extracted_flags[i+1]:
                    current_group.append(sections[i+1])
                    extracted_flags[i+1] = True
            if current_group:
                extracted_groups.append(current_group)
    return extracted_groups


migrant['sections_with_keyword'] = migrant['text'].apply(lambda x: extract_sections(x))

def flatten_to_strings(list_of_lists):
    """ 
    the extracted lists are being joined to a list
    """
    return [' '.join(sublist) for sublist in list_of_lists]


migrant['sections'] = migrant['sections_with_keyword'].apply(flatten_to_strings)


### Filter sections with matches and drop unneccesary columns

migrant = migrant[migrant['sections'].apply(lambda x: len(x) > 0)]
migrant_masked = migrant[['_id', 'sections', 'outlet']]

### Mask matches

def replace_words(text_list, words_to_replace, replacement_word):
    new_text_list = []
    for text in text_list:
        for old_word in words_to_replace:
            # Using regular expression to match whole words
            pattern = r'\b{}\b'.format(re.escape(old_word))
            text = re.sub(pattern, replacement_word, text, flags=re.IGNORECASE)
        new_text_list.append(text)
    return new_text_list

words_to_replace = uniquematch
replacement_word = "[Zielgruppe]"
print(uniquematch)

df['masked'] = df['response'].apply(lambda x: replace_words(x, words_to_replace, replacement_word))

### Explode masked column

migrant_masked = migrant_masked.explode('masked')
#Reset index
migrant_masked.reset_index(drop=True, inplace=True)

### Check for more Matches

re_ukraine = '[Zielgruppe]'
re_türkei = '[Gruppe2]'
re_syrien = '[Gruppe3]'
re_somalien = '[Gruppe4]'
re_afghanistan = '[Gruppe5]'
re_flüchtling = '[Gruppe6]'
re_einwanderer = '[Gruppe6]'

ukr_pattern = r'\b\w*ukraine\w*\b' #
ukr1_pattern = r'\b\w*ukrainisch\w*\b' #
tur_pattern = r'\b\w*türkei\w*\b' #
tur1_pattern = r'\b\w*türki\w*\b' #
syr_pattern = r'\b\w*syrer\w*\b' 
syr1_pattern = r'\b\w*syri\w*\b' #
som_pattern = r'\b\w*somali\w*\b' 
afg_pattern = r'\b\w*afghan\w*\b'
fl_pattern = r'\b\w*flüchtl\w*\b' #
fl1_pattern = r'\w*geflücht\w*\b'
ein_pattern = r'\w*einwanderer\w*\b'
ein1_pattern = r'\w*zuwanderer\w*\b'
ein2_pattern = r'\w*immigrant\w*\b'
ein3_pattern = r'\w*migrant\w*\b'



patterns = [ukr_pattern, ukr1_pattern, tur_pattern, tur1_pattern, syr_pattern, syr1_pattern, som_pattern, afg_pattern, fl_pattern, fl1_pattern, ein_pattern, ein1_pattern, ein2_pattern, ein3_pattern]

# Function to find which patterns are found in the text
def find_patterns(text, patterns):
    found_patterns = set()  
    for pattern in patterns:
        if re.findall(pattern, text,re.IGNORECASE):
            found_patterns.add(pattern)
    return found_patterns

all_found_patterns = set()

# Iterate over each row in the DataFrame column and update the set of found patterns
for t in df['masked']:
    all_found_patterns.update(find_patterns(t, patterns))

# Convert the set to a list 
all_found_patterns_list = list(all_found_patterns)
print(all_found_patterns_list)



data = df

data['masked'] = data['text'].apply(replace_ukraine_special)
data['masked'] = data['text'].apply(replace_ukrainisch_special)

data['masked'] = data['text'].apply(replace_türkisch_special)
data['masked'] = data['text'].apply(replace_türke_special)

data['masked'] = data['text'].apply(replace_syrien_special)
data['masked'] = data['text'].apply(replace_syrisch_special)

data['masked'] = data['text'].apply(replace_somalisch_special)
data['masked'] = data['text'].apply(replace_somali_special)

data['masked'] = data['text'].apply(replace_afghanistan_special)
data['masked'] = data['text'].apply(replace_afghanisch_special)

data['masked'] = data['text'].apply(replace_flüchtl_special)
data['masked'] = data['text'].apply(replace_geflüchte_special)

data['masked'] = data['text'].apply(replace_migrant_special)
data['masked'] = data['text'].apply(replace_einwa_special)
data['masked'] = data['text'].apply(replace_zuwan_special)

def replace_migrant_special(text, exclude_sequences=['abc']):
    pattern = r'migrant(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        
        if len(following_chars) <= 2:  
            return re_einwanderer
        else:  
            return re_einwanderer + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_einwa_special(text, exclude_sequences=['abc']):
    pattern = r'einwanderer(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1) 
        
    
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_einwanderer
        else:  
            return re_einwanderer + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_zuwan_special(text, exclude_sequences=['abc']):
    pattern = r'zuwanderer(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1) 
        
    
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_einwanderer
        else:  
            return re_einwanderer + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_flüchtl_special(text, exclude_sequences=['abc']):
    pattern = r'flüchtling(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1) 
    
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_flüchtling
        else: 
            return re_flüchtling + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_geflüchte_special(text, exclude_sequences=['t']):
    pattern = r'(?<!hin)Geflüchte(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        
    
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2: 
            return re_flüchtling
        else:  
            return re_flüchtling + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_afghanistan_special(text, exclude_sequences=['abc']):
    pattern = r'afghanistan(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2: 
            return re_afghanistan
        else: 
            return re_afghanistan + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_afghanisch_special(text, exclude_sequences=['abc']):
    pattern = r'afghanisch(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_afghanistan
        else:  
            return re_afghanistan + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_somalisch_special(text, exclude_sequences=['abc']):
    pattern = r'somalisch(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_somalien
        else: 
            return re_somalien + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_somali_special(text, exclude_sequences=['abc']):
    pattern = r'somali(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2: 
            return re_somalien
        else:  
            return re_somalien + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_syrien_special(text, exclude_sequences=['abc']):
    pattern = r'syri(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_syrien
        else:  
            return re_syrien + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_syrisch_special(text, exclude_sequences=['abc']):
    pattern = r'(?<!weiß)(?<!bela)syrisch(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 2:  
            return re_syrien
        else:  
            return re_syrien + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_türke_special(text, exclude_sequences=['abc']):
    pattern = r'(?<!bela)türke(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 5:  
            return re_türkei
        else:  
            return re_türkei + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_türkisch_special(text, exclude_sequences=['abc']):
    pattern = r'(?<!bela)türki(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 5:  
            return re_türkei
        else: 
            return re_türkei + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_ukraine_special(text, exclude_sequences=['abc']):
    pattern = r'(?<!weiß)(?<!bela)ukraine(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 5: 
            return re_ukraine
        else:  
            return re_ukraine + following_chars
    
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_ukrainisch_special(text, exclude_sequences=['abc']):
    pattern = r'(?<!weiß)(?<!bela)ukrainisch(\w*)\b'
    
    def replace(match):
        following_chars = match.group(1)  
        if following_chars.lower() in exclude_sequences:
            return match.group(0)
        
        if len(following_chars) <= 5:  
            return re_ukraine
        else:  
            return re_ukraine + following_chars
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Replace

data['masked'] = data['masked'].apply(replace_ukraine)
data['masked'] = data['masked'].apply(replace_ukrainisch)

data['masked'] = data['masked'].apply(replace_türkei)
data['masked'] = data['masked'].apply(replace_türkisch)
data['masked'] = data['masked'].apply(replace_türke)
data['masked'] = data['masked'].apply(replace_türkin)

data['masked'] = data['masked'].apply(replace_syrien)
data['masked'] = data['masked'].apply(replace_syrisch)
data['masked'] = data['masked'].apply(replace_syrer)

data['masked'] = data['masked'].apply(replace_somalia)

data['masked'] = data['masked'].apply(replace_afghanistan)
data['masked'] = data['masked'].apply(replace_afghanisch)
data['masked'] = data['masked'].apply(replace_afghane)

data['masked'] = data['masked'].apply(replace_flüchtling)
data['masked'] = data['masked'].apply(replace_geflüchtete)

data['masked'] = data['masked'].apply(replace_einwanderer)
data['masked'] = data['masked'].apply(replace_einwanderin)
data['masked'] = data['masked'].apply(replace_migrant)
data['masked'] = data['masked'].apply(replace_immigrant)
data['masked'] = data['masked'].apply(replace_zuwanderer)
data['masked'] = data['masked'].apply(replace_zuwanderin)

### Remove more matches (Ukraine)

def replace_ukraine(text):
    pattern = r'ukraine([a-zA-Z]{0,3})\b|ukraine(\w*)'
    
    def replace(match):
        if match.group(1): 
            return re_ukraine
        else:  
            return re_ukraine + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_ukrainisch(text):
    pattern = r'ukrainisch(\w{0,2})\b|ukrainisch(\w*)'
    
    def replace(match):
        if match.group(1): 
            return re_ukraine
        else:  
            return re_ukraine + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Remove more matches (Türkei)

def replace_türkei(text):
    pattern = r'\bTürkei([a-zA-Z]{0,1})\b|\bTürkei(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_türkei
        else:  
            return re_türkei + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_türkisch(text):
    pattern = r'\bTürkisch(\w{0,2})\b|\bTürkisch(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_türkei
        else:  
            return re_türkei + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_türke(text):
    pattern = r'\bTürke(\w{0,2})\b|\bTürki(\w*)|\bTürken(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_türkei
        else:  
            return re_türkei + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_türkin(text):
    pattern = r'\bTürki(\w{0,2})\b|\bTürkin(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_türkei
        else: 
            return re_türkei + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Remove more matches (Syrien)

def replace_syrien(text):
    pattern = r'syrien(\w{0,2})\b|syrien(\w*)'
    
    def replace(match):
        if match.group(1): 
            return re_syrien
        else:  
            return re_syrien + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_syrisch(text):
    pattern = r'syrisch(\w{0,2})\b|syrisch(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_syrien
        else:  
            return re_syrien + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_syrer(text):
    pattern = r'syrer(\w{0,5})\b|syrer(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_syrien
        else:  
            return re_syrien + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Remove more matches (Somalia)

def replace_somalia(text):
    pattern = r'somali(\w*)'
    def replace(match):
        if match.group():  
            return re_somalien
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Remove more matches (Afghansitan)

def replace_afghanistan(text):
    pattern = r'\bAfghanistan(\w{0,2})\b|\bAfghanistan(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_afghanistan
        else:  
            return re_afghanistan + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_afghanisch(text):
    pattern = r'afghanisch(\w{0,2})\b|afghanisch(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_afghanistan
        else:  
            return re_afghanistan + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_afghane(text):
    pattern = r'\bafghan(\w*)'
    def replace(match):
        if match.group(): 
            return re_afghanistan
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Remove more matches (Flüchtlinge)

def replace_flüchtling(text):
    pattern = r'flüchtling([a-zA-Z]{0,2})\b|flüchtlings(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_flüchtling
        else:  
            return re_flüchtling + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_geflüchtete(text):
    pattern = r'\bgeflüchtete([a-zA-Z]{0,2})\b|\bgeflüchtete(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_flüchtling
        else:  
            return re_flüchtling + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Remove more matches (Migr)

def replace_migrant(text):
    pattern = r'\bmigrant([a-zA-Z]{0,5})\b|\bmigrant(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_einwanderer
        else:  
            return re_einwanderer + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_immigrant(text):
    pattern = r'\bimmigrant([a-zA-Z]{0,5})\b|\bimmigrant(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_einwanderer
        else: 
            return re_einwanderer + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_einwanderer(text):
    pattern = r'einwanderer([a-zA-Z]{0,1})\b|einwanderer(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_einwanderer
        else:  
            return re_einwanderer + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_einwanderin(text):
    pattern = r'\beinwanderin([a-zA-Z]{0,3})\b|\beinwanderin(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_einwanderer
        else:  
            return re_einwanderer + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_zuwanderer(text):
    pattern = r'\bzuwanderer([a-zA-Z]{0,1})\b|\bzuwanderer(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_einwanderer
        else:  
            return re_einwanderer + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

def replace_zuwanderin(text):
    pattern = r'\bzuwandererin([a-zA-Z]{0,3})\b|\bzuwandererin(\w*)'
    
    def replace(match):
        if match.group(1):  
            return re_einwanderer
        else:  
            return re_einwanderer + (match.group(2) if match.group(2) else '')
    
    return re.sub(pattern, replace, text, flags=re.IGNORECASE)

### Labels

# Define mapping of keywords to labels
keyword_label_mapping = {
    '[Gruppe2]': 'Somalia',
    '[Gruppe7]': 'Afghanistan',
    '[Gruppe4]': 'Flüchtlinge',
    '[Gruppe1]': 'Migranten',
    '[Gruppe3]': 'Ukraine',
    '[Gruppe5]': 'Türkei',
    '[Gruppe6]': 'Syrien',

}

def get_labels(text):
    labels = []
    for keyword, label in keyword_label_mapping.items():
        if keyword in text:
            labels.append(label)
    return ', '.join(labels) if labels else 'None'

migrant_masked['labels'] = migrant_masked['masked'].apply(get_labels)

### To csv

migrant_masked.to_csv('../Masked Data/migrant_masked.csv')
migrant.to_csv('../ExtractedData/migrant.csv')
