==> Code/03_Sampling.ipynb <==
import pandas as pd
import re

mig = pd.read_csv('../Masked Data/migrant_masked.csv')
ukr = pd.read_csv('../Masked Data/ukraine_masked.csv')
tur = pd.read_csv('../Masked Data/türkei_masked.csv')
syr = pd.read_csv('../Masked Data/syrien_masked.csv')
som = pd.read_csv('../Masked Data/somalia_masked.csv')
afg = pd.read_csv('../Masked Data/afghanistan_masked.csv')
flu = pd.read_csv('../Masked Data/flüchtling_masked.csv')

df1 = mig[mig.outlet == 'national'].sample(4)
df2 = mig[mig.outlet == 'regional'].sample(4)
df3 = mig[mig.outlet == 'right'].sample(4)
df4 = mig[mig.outlet == 'left'].sample(3)

df5 = ukr[ukr.outlet == 'national'].sample(4)
df6 = ukr[ukr.outlet == 'regional'].sample(4)
df7 = ukr[ukr.outlet == 'right'].sample(3)
df8 = ukr[ukr.outlet == 'left'].sample(3)

df9 = tur[tur.outlet == 'national'].sample(4)
df10 = tur[tur.outlet == 'regional'].sample(4)
df11 = tur[tur.outlet == 'right'].sample(3)
df12 = tur[tur.outlet == 'left'].sample(3)

df13 = syr[syr.outlet == 'national'].sample(4)
df14 = syr[syr.outlet == 'regional'].sample(4)
df15 = syr[syr.outlet == 'right'].sample(3)
df16 = syr[syr.outlet == 'left'].sample(3)

df17 = flu[flu.outlet == 'national'].sample(4)
df18 = flu[flu.outlet == 'regional'].sample(4)
df19 = flu[flu.outlet == 'right'].sample(4)
df20 = flu[flu.outlet == 'left'].sample(3)

df21 = som[som.outlet == 'national'].sample(5)
df22 = som[som.outlet == 'regional'].sample(5)
df23 = som[som.outlet == 'right'].sample(4)
df24 = som[som.outlet == 'left'].sample(0)

df25 = afg[afg.outlet == 'national'].sample(4)
df26 = afg[afg.outlet == 'regional'].sample(4)
df27 = afg[afg.outlet == 'right'].sample(3)
df28 = afg[afg.outlet == 'left'].sample(3)

sample3 = pd.concat([df1, df2, df3, df4, df5, df6, df7, df8, df9, df10, df11, df12, df13, df14, df15, df16, df17, df18, df19, df20, df21, df22, df23, df24, df25, df26, df27, df28])
sample3 =sample3.sample(frac=1).reset_index(drop=True)
sample3.to_csv('../Samples/Full Samples/sample3_full.csv')

sample17.shape

sample3_filter = sample3[['_id', 'masked']]
sample3_filter.to_csv('../Samples/sample3.csv')

ids1 = sample1['Unnamed: 0'].unique().tolist()
ids2 = sample2['Unnamed: 0'].unique().tolist()
ids3 = sample3['Unnamed: 0'].unique().tolist()
ids4 = sample4['Unnamed: 0'].unique().tolist()
ids5 = sample5['Unnamed: 0'].unique().tolist()
ids6 = sample6['Unnamed: 0'].unique().tolist()
ids7 = sample7['Unnamed: 0'].unique().tolist()
ids8 = sample8['Unnamed: 0'].unique().tolist()
ids9 = sample9['Unnamed: 0'].unique().tolist()
ids10 = sample10['Unnamed: 0'].unique().tolist()
ids11 = sample11['Unnamed: 0'].unique().tolist()
ids12 = sample12['Unnamed: 0'].unique().tolist()
ids13 = sample13['Unnamed: 0'].unique().tolist()
ids14 = sample14['Unnamed: 0'].unique().tolist()
ids15 = sample15['Unnamed: 0'].unique().tolist()
ids16 = sample16['Unnamed: 0'].unique().tolist()
ids17 = sample17['Unnamed: 0'].unique().tolist()
ids18 = sample18['Unnamed: 0'].unique().tolist()
ids19 = sample19['Unnamed: 0'].unique().tolist()
ids20 = sample20['Unnamed: 0'].unique().tolist()

ids = [60302, 3136, 20170, 63078, 56507, 4965, 474, 6912, 11448,228, 56760, 2476, 575, 6372, 1119, 83, 57139, 18185, 56856, 1003, 2695, 1501, 1270, 22263, 532, 430, 62391, 3037]
ids_to_remove= ids + ids1 + ids2 + ids3 + ids4 +ids5 + ids6 +ids7 +ids8 +ids9 +ids10 +ids11 +ids12 +ids13 +ids14 +ids15 +ids16 +ids17 +ids18 +ids19 +ids20

mig = mig[~mig['Unnamed: 0'].isin(ids_to_remove)]
ukr = ukr[~ukr['Unnamed: 0'].isin(ids_to_remove)]
tur = tur[~tur['Unnamed: 0'].isin(ids_to_remove)]
syr = syr[~syr['Unnamed: 0'].isin(ids_to_remove)]
som = som[~som['Unnamed: 0'].isin(ids_to_remove)]
afg = afg[~afg['Unnamed: 0'].isin(ids_to_remove)]
flu = flu[~flu['Unnamed: 0'].isin(ids_to_remove)]


# ids1 = sample1['_id'].unique().tolist()
# ids2 = sample2['_id'].unique().tolist()
# ids3 = sample3['_id'].unique().tolist()
# ids4 = sample4['_id'].unique().tolist()
# ids5 = sample5['_id'].unique().tolist()
# ids6 = sample6['_id'].unique().tolist()
# ids7 = sample7['_id'].unique().tolist()
# ids8 = sample8['_id'].unique().tolist()
# ids9 = sample9['_id'].unique().tolist()

ids = ['6327128f3dd49b919c121751', '632712903dd49b919c12cb35', '632712913dd49b919c1347a1']
ids_to_remove= ids #+ ids1 #+ ids2 #+ ids3 #+ ids4 #+ids5 #+ ids6 #+ids7 #+ids8 #+ids9

mig = mig[~mig['_id'].isin(ids_to_remove)]
ukr = ukr[~ukr['_id'].isin(ids_to_remove)]
tur = tur[~tur['_id'].isin(ids_to_remove)]
syr = syr[~syr['_id'].isin(ids_to_remove)]
#ita = ita[~ita['_id'].isin(ids_to_remove)]
som = som[~som['_id'].isin(ids_to_remove)]
afg = afg[~afg['_id'].isin(ids_to_remove)]
flu = flu[~flu['_id'].isin(ids_to_remove)]

sample20.to_csv('../Samples/Full Samples/sample20_full.csv')
