==> Code/04_gpt.ipynb <==
import openai

openai.api_key = ''


import pandas as pd
import re
import random

mylist = [
'- Ukraine-Konflikt, Polen, Eindrücke aus Przemysl: Ukrainer erreichen den Bahnhof in Krakau. 7. März 2022. Im Krieg in der Ukraine sind nach Angaben des ukrainischen Verteidigungsministers Oleksij Resnikow bislang mindestens 38 Kinder durch russische Angriffe zu Tode gekommen. Insgesamt seien seit Kriegsbeginn Russlands über 400 Zivilisten getötet worden, so der Verteidigungsminister am Dienstag. Die Zahlen ließen sich indes nicht unabhängig überprüfen.', 'Insgesamt wurden seinen Angaben zufolge 400 Zivilisten getötet und mindestens 800 verletzt. Diese Daten seien jedoch „definitiv unvollständig“, sagte Resnikow. Die Zahlen unabhängig zu bestätigen, war zunächst nicht möglich. Zuletzt verifizierte das UN-Hochkommissariat für Menschenrechte 364 Tote Ukrainer im Zusammenhang russischer Angriffe, darunter 25 Kinder. Die tatsächlichen Zahlen lägen jedoch um einiges höher, vermutet die UN.',
'- Die Bundesregierung hat angedeutet, man werde ein mögliches Gesuch der spanischen Regierung zur Lieferung deutscher Leopard-Panzer an die Ukraine wohlwollend prüfen. Aber Panzerlieferungen könnten auch viel einfacher auf direktem Wege aus Deutschland in die Ukraine erfolgen, so Melnyk - wenn denn die Bundesregierung mitspiele: "Die deutsche Industrie ist bereit, diese Waffensysteme sofort zu liefern. Sofort. Ohne Ringtäusche. Ohne Manöver. Ohne Umwege. Und wir können nicht verstehen, was auf dem Wege noch geschehen muss, wieso kein grünes Licht gegeben wird von der Ampelkoalition, oder vom Sicherheitsrat, oder vom Kanzleramt - denn davon hängt ab, wie viele Ukrainer heute und morgen und übermorgen sterben müssen."',
'- Nie war die Ukraine so gefährdet wie jetzt im Krieg mit Russland. Von internationalen Gratulanten wird vor allem erwartet, schwere Waffen zu liefern. Eine Ukrainerin steht auf einem zerstörten russischen Panzer Die Ukraine begeht heute ihren 31. Unabhängigkeitstag.', '\nKiew Im Schatten des seit einem halben Jahr andauernden russischen Angriffskriegs begeht die Ukraine am Mittwoch ihren Nationalfeiertag. Der Unabhängigkeitstag sei ein wichtiges Datum für die Ukrainer und Ukrainerinnen „und damit leider auch für unseren Feind“, sagte Präsident Wolodimir Selenski. In Kiew herrschte Nervosität wegen befürchteter russischer Raketenangriffe auf große Städte. Selenskyj rief die Bevölkerung auf, sich an die Ausgangssperren zu halten und bei Luftalarm in Sicherheit zu bringen. „Denkt daran, wir sollen alle gemeinsam den Sieg erleben“, sagte er.', 'Zum Unabhängigkeitstag wurde in Kiew erwartet, dass noch weitere Länder neue Rüstungszusagen machen. Zu dem doppelt bedeutsamen Datum sendet die ARD ihre „Tagesthemen“ live (22.15 Uhr) aus Kiew, moderiert von Caren Miosga. Die Moderatorin habe in den vergangenen Tagen mit Ukrainerinnen und Ukrainern über deren Leben und das Überleben im Krieg gesprochen, teilte der NDR mit. Drehorte der Reportagereise waren Lwiw und Tscherniwzi sowie Kiew mit seinen Vororten Butscha, Moschtschun und Irpin.\n',
'- Wir unterstützen die Ukrainer bei ihrem Freiheitskampf gegen den Aggressor Putin. Wir helfen Flüchtlingen, besonders den vielen Frauen und Kindern, wenn sie zu uns nach Deutschland kommen. Wir sind für die Sanktionen gegen die russische Finanz-Elite und die Oligarchen, die Putin unterstützen. Niemand darf das unendliche Leid der Ukrainer gegen finanzielle Sorgen bei uns ausspielen. Aber: Auch die Bürger in Deutschland haben jetzt Hilfe nötig, damit die Folgen dieses Krieges sie nicht überfordern.',
'- Musik und Lieder spielen in der Ukraine eine große Bedeutung – insbesondere für das Nationalbewußtsein. Einer der bekanntesten Rockmusiker der Ukraine, Oleg Skrypka, unterstreicht dies im Interview mit der JUNGEN FREIHEIT. Skrypka wurde 1964 in Ghafurow (Tadschikistan) geboren. Sein Vater ist Ukrainer, seine Mutter Russin. Nach dem Abitur gründete er noch zu Sowjetzeiten 1986 in Kiew die Rockband „Vopli Vidopliassova“. Die Musik ist eine Mischung aus Rock’n’Roll und Ethno-Folk, vereint Elemente aus Punk, Heavy Metal und Elektro-Pop. Mit seinem Sohn Ustym nahm er 2021 ein 100 Jahre altes patriotisches Lied („Wir wachsen“) auf, das der Pfadfinder-Bewegung gewidmet ist. Das Musikvideo dazu hat seit Beginn des russischen Angriffs auf die Ukraine am 24. Februar 2022 enorme virale Verbreitung bekommen.', 'Bei diesen Songs war das anders. Ich denke, es ist normal für Künstler, ein besonderes Gefühl für die Zukunft zu haben. Mir jedenfalls standen viele der Probleme schon damals vor Augen, die wir heute mit Rußland haben. Deshalb habe ich all diese patriotischen Lieder eingespielt. „Jetzt ist große Solidarität unter Ukrainern wachgeworden“ Wie reagieren Ihre Musikerkollegen in der Ukraine auf den russischen Einmarsch? Helfen sie mit ihrer Kunst? Fliehen sie? Oder greifen sie zu den Waffen?', 'Skrypka: Momentan stehen wir alle noch unter Schock. Natürlich versuchen wir erst mal nur, uns zu verteidigen. Wir hoffen sehr, daß wir diesen Krieg möglichst schnell gewinnen. Erst danach stellt sich die Frage, was mit russischsprachigen Künstlern in der Ukraine passiert. Die meisten werden nach dem Krieg einfach auf ukrainisch weitermachen. Was mit der Minderheit passiert, die es sich nicht anders überlegen will, weiß ich nicht. Vielleicht gehen die nach Europa. Eigentlich glaube ich aber nicht, daß die Frage der Sprache in der Ukraine wirklich so wichtig ist. Das wurde eher von der russischen Propaganda aufgebauscht. „Ukrainer schöpfen ihr Nationalbewußtsein aus ihrer Musik“ Welche Rolle spielen Musik und Lieder im ukrainischen Selbstverständnis?', 'Skrypka: Mir gefällt diese Frage sehr! Für Ukrainer war ihr Liedgut immer schon besonders wichtig. Über viele Jahrhunderte hatten die Ukrainer keinen eigenen Staat. Also mußten sie ihr nationales Selbstverständnis aus ihrer Musik schöpfen. Deshalb haben wir Ukrainer auch eine ganze Reihe traditioneller und moderner Volkslieder.', 'Skrypka: Unsere Geschichte und unsere Kultur reichen sehr weit zurück. In der Ukraine sind sich viele Menschen dessen nicht bewußt, aber wir leben in dieser Tradition. Daß unsere Kultur so tief in unserem Leben verwurzelt ist, hat sie letztendlich auch gerettet. Auch durch die 300 Jahre hindurch, in der wir Ukrainer keinen eigenen Staat hatten und von Rußland besetzt waren. Unsere Tradition hat uns mit unserem Land verbunden, ähnlich wie bei den Mexikanern, Chinesen und Indern. Diese Völker hatten aber ihren eigenen Staat, wir nicht. Wenn wir diesen Krieg gewinnen sollten, wäre das für uns also ein kompletter Neuanfang.', 'Skrypka: Die ganze Zeit der russischen Besatzung über – etwa 300 Jahre lang – gab es ukrainische Dissidenten. Das ist auch so eine Tradition von uns Ukrainern, eine Idee von unserem Land im Untergrund zu pflegen. Diese Idee bricht jetzt mit aller Kraft durch, daß wir endlich einen ukrainischen Staat bekommen.', 'Skrypka: Ukrainer wünschen sich von ihrer Zukunft, daß sie in einem modernen Land leben können, in dem sie frei sind. Sie träumen von einer Gesellschaft, die ihnen die Möglichkeit gibt, ihre Potentiale auszuschöpfen. Das Problem mit Rußland hat dabei nichts mit Sprache oder Nationalität zu tun. Ich habe russische Freunde aus Moskau, die trotzdem großartige ukrainische Patrioten sind. Die Frage ist doch, ob man dieses Land und diese Nation liebt. Und ich denke sogar, daß in der Zukunft viele Russen in die Ukraine kommen werden. Solange sie gerne hier leben und unserem Land etwas Gutes tun, habe ich absolut kein Problem damit.', 'Skrypka: Das ist eine heikle Frage (lacht). Ich weiß nicht, ich bin kein Politiker. Die Welt verändert sich. Während der ersten Revolution 2004 und nach 2014 haben die Ukrainer eine große Sehnsucht nach der europäischen Gemeinschaft verspürt. Es gibt jetzt die reale Möglichkeit für uns, in diese europäische Familie aufgenommen zu werden. Das ist sehr gut. Wir sehen aber auch, daß das alles nicht so einfach ist. Die Staaten der EU müssen sich erst einmal untereinander darauf einigen, uns eine Chance zu geben. Wir sind da nicht mehr so naiv wie in der Vergangenheit. Trotzdem wollen wir natürlich ein gutes Verhältnis zu allen europäischen Ländern haben. Das ist sehr wichtig für die Ukraine.', '„Dieser Krieg ist keine Privatsache zwischen Ukrainern und Russen“ Führt der Krieg in der Ukraine zu einer „zweiten Geburt einer Nation“, wie es Historiker nennen? Rückt das Land unter den Angriffen erst jetzt richtig zusammen? Erfahren viele Ukrainer im Moment wirklich, was es heißt, Teil einer Nation zu sein?', 'Skrypka: Sie haben schon die Antwort selbst gegeben (lacht noch lauter)! Wir Ukrainer erleben tatsächlich gerade die Wiedergeburt unserer Nation. Es fällt mir zwar schwer, das zu sagen, aber in gewisser Weise – so bitter das klingt – haben wir diesen Krieg gebraucht. Es ist, als ob Gott uns diese Katastrophe geschenkt hat, damit wir Ukrainer endlich Ukrainer werden. Ich möchte aber auch sagen, daß dieser Krieg nicht mehr einfach eine Privatsache zwischen Ukrainern und Russen ist. Wir fechten diesen Kampf gemeinsam mit Europäern, Japanern und Amerikanern aus. Die Russen leben immer noch in der Vergangenheit. Sie müssen ihre aggressive Art zu denken und zu handeln endlich aufgeben. Das ist einfach nicht menschlich. Und vielleicht geht es am Ende nicht nur um die Wiedergeburt der ukrainischen Nation, sondern um die Geburt einer neuen Gesellschaftsordnung weltweit.',
'- Die USA wollen als Teil eines neuen Militärhilfe-Pakets in Höhe von 1,85 Milliarden US-Dollar (1,74 Milliarden Euro) ein Patriot-Luftabwehrsystem in die Ukraine schicken, auch neue Munition wurde den Ukrainern versprochen. +++ Alle Entwicklungen zum Krieg gegen die Ukraine im Liveblog +++',
'- Szenen, die viele Ukrainerinnen und Ukrainer in Deutschland von ihren Familien und Freunden ebenso geschildert bekommen. „Ich bin heute morgen um 6 Uhr aufgewacht und habe sofort meine Familie in Kiew angerufen, weil ich ein ungutes Gefühl hatte“, erzählt Olena Shutovska. Sie und ihr Mann leben in Berlin, sind an diesem Vormittag zum Brandenburger Tor gekommen, um gemeinsam mit Hunderten anderen für die Unterstützung der Ukraine zu demonstrieren. „Wir mussten irgendetwas tun. Wir fühlen uns machtlos, können nicht helfen“, sagt die 29-Jährige.', 'Entsetzen, Angst und Wut sind die vorherrschenden Emotionen unter den Demonstranten. „Ihr seid alle in Gefahr. Ihr steht hier nicht nur für die Ukraine, sondern auch für Europa“, schallt durch ein Megafon. Drei junge Ukrainerinnen, die die Versammlung mit organisiert haben, verkünden ihre Statements nacheinander auf Ukrainisch, Englisch und Deutsch. Auch die ehemalige Grünen-Abgeordnete Marieluise Beck ergreift das Wort: „Es müssten hier auch mehr Deutsche stehen. Aber vielleicht kommt das noch, wenn sie anfangen zu verstehen.“', 'Auf Facebook vernetzen sich indes in Deutschland lebende Ukrainerinnen und Ukrainer, um zu helfen. In der Gruppe „Ukrainer in Berlin“ wird seit dem Morgen immer wieder nach Transportmöglichkeiten gefragt, ob Leute in die Ukraine fahren, ob man Familien herausbringen kann. Unter zahlreichen Beiträgen werden mazedonische, serbische, slowenische und polnische Handynummern gepostet, versehen mit Informationen, wie „kann morgen abgeholt werden“ oder „Igor heute Abend oder morgen“.',
'- Er vergleicht: „Napoleon, Hitler und Stalin mussten alle ihre Armeen angesichts des Winters in der Steppe in Bewegung halten, und jetzt – seine Invasion ist auf dem Rückzug – verschanzt Wladimir Putin seine Streitkräfte für den Winter, um auf eine neue russische Offensive im Frühjahr zu warten.“ Die Ukrainer seien aber „besser gerüstet und motiviert“ und würden den Druck im Donbass aufrechterhalten, glaubt der Analyst. Gut gerüstet: Ein ukrainischer Soldat posiert Winter-Unterstand an der Front bei Saporischschja. © IMAGO/Celestino Arce\nIm Kampf mit Russland: Ex-US-General glaubt an militärischen Sieg der Ukraine',
'- Bei seinen beiden Besuchen in Kiew sei Johnson mit ihm durch die Straßen gelaufen und habe auch normale Ukrainerinnen und Ukrainer kennengelernt, schrieb Selenskyj. Wir haben sogar scherzhaft angefangen, ihn ‚Boris Johnsoniuk‘ zu nennen, was wie ein ukrainischer Nachname klingt...',
'- Zuletzt hatten russische Raketen- und Drohnenangriffe erhebliche Teile der ukrainischen Strom- und Wasserinfrastruktur zerstört. Mehrere Millionen Ukrainer haben seitdem jeden Tag nur stundenweise Strom. Vor diesem Hintergrund lieferte auch Deutschland im Oktober das Luftabwehrsystem Iris-T.' 
]



def query_LLM(sample):

    system_prompt = """
    Du bist ein Journalist und schreibst Zeitungsartikel (ohne Überschrift). Du bekommst ein Thema vorgegeben und musst dazu entsprechende Zeitungsartikel schreiben. Außerdem wird dir eine Anzahl an Artikel und Länge vorgegeben. Du musst sicherstellen, dass deine Artikel diese Vorgaben erfüllen.
    """

    user_prompt = """
    Schreibe 10 Zeitungsartikel zum Thema 'Ukrainer'. 
    Die Längenvorgabe ist: 120 Tokens pro Zeitungsartikel.
    Bitte sei kreativ und schreibe einzigartige Artikel. 

    Hier sind einige Beispiele:
    {sample}

    Du sollst die von mir gegebene Beispiel nicht einfach abändern oder umschreiben.
    """

    response = openai.ChatCompletion.create(
        model="gpt-3.5-turbo",
        temperature=0.7,
        # top_p=0.9,
        # frequency_penalty=0,
        # presence_penalty=0,  
        messages=[
            {"role": "system", "content": system_prompt},
            {"role":   "user", "content": user_prompt}
            ]
    )

    clippings = []
    for choice in response['choices']:
        clipping_text = choice['message']['content']
        clippings.append(clipping_text)

    return response

responses = list()

for i in range(8):
    result = ""
    sam = random.sample(mylist, 4)

    for i in sam:
        result += i + '\n' 

    generated_clippings = query_LLM(result)
    responses.append(generated_clippings['choices'][0]['message']['content'])

responses

s = ','.join(responses)
s

seg = re.split(r'\d+\.\s*', s)
#seg = re.split(r'(?m)^-\s+', generated_clippings['choices'][0]['message']['content'])
# seg = re.split(r'\n', generated_clippings['choices'][0]['message']['content'])

# def clean_segment(segment):
#     match = re.search(r'\"(.*)\"', segment)
#     if match:
#         return match.group(1)
#     return ''

# words = [clean_segment(segment) for segment in seg if segment.strip()]

seg = [x for x in seg if x != '']

print(seg)

df = pd.DataFrame(seg, columns=['Texts'])
df

df.to_csv('../Ergebnisse/Generated/gpt/ukr_v1.csv')

generated_clippings
