==> Code/08_Clean_GeneratedData.ipynb <==
import pandas as pd
import re

df = pd.read_csv('../Ergebnisse/Generated/leolm/round2/Pos/ukr_pos_r2.csv')
df.shape

Remove short rows because those aren't articles

df['length'] = df['response'].apply(len)
df[df['length']<100]


# Filter the DataFrame to keep only rows with text length >= min_length
df = df[df['length'] >= 100].reset_index(drop=True)

# Drop the 'length' column if it's no longer needed
df = df.drop(columns=['length'])

Remove Numbers at the start

# pattern = r'(^\d+[)\.] )|(^(Zeitungsartikel ?\w{1,3}:\s*))|(^(?:\*\* ?Artikel\s*\d*\s*[-:] \s*))'
# pattern = r'(^(Im fünften und letzten Artikel mit dem Titel\s))|(^(Zeitungsartikel \d+\s-\s))|(^(Artikel\s\d:\n))'
pattern = r'(^\w{1,3}\s*--\s*)|(^\d+[)\.] )|(^(Zeitungsartikel\s*\d*\s*[:-]\s*))|(^(Artikel\s*\d*\s*[-:]\s*))'

df['response'] = df['response'].str.replace(pattern, '', regex=True, case=False)

df = df.drop([0])

df.iloc[30:60]
# df.iloc[0,1]

df.to_csv('../Ergebnisse/Generated/leolm/round2/Pos/ukr_pos_clean.csv')
